chore: 메인 페이지에서 하이드레이션이 아닌 fetch 를 사용해서 포스트 데이터를 가져오도록 수정
